requirements.txt
flask
tensorflow
keras
Pillow
waitress
